Added Interfaces for Withdraw, Passive and Deposit. Adjusted quantile filtering of AggregationService

+++ firecasting/common-domain/src/main/java/dk/gormkrings/inflation/IInflationFactory.java
@@ -2,5 +2,5 @@ package dk.gormkrings.inflation;
 
 public interface IInflationFactory {
     IInflation createInflation();
-    IInflation createInflation(float inflationPercentage);
+    IInflation createInflation(double inflationPercentage);
 }
